+++ assets/js/app.js
@@ -21,15 +21,15 @@
 // Include phoenix_html to handle method=PUT/DELETE in forms and buttons.
 import "phoenix_html"
 // Establish Phoenix Socket and LiveView configuration.
-import { Socket } from "phoenix"
-import { LiveSocket } from "phoenix_live_view"
+import {Socket} from "phoenix"
+import {LiveSocket} from "phoenix_live_view"
 import topbar from "../vendor/topbar"
 import AudioFile from "./hooks/audio-file"
 import DraggableCheckpoints from "./hooks/draggable-checkpoints"
 import MediaFileInput from "./hooks/media-file-input"
 import SortableUploadEntries from "./hooks/sortable-upload-entries"
 import UploadInProgress from "./hooks/upload-in-progress"
-import { UpChunkUploader, uploadEventHandlers } from "./upchunk_upload"
+import {UpChunkUploader, uploadEventHandlers} from "./upchunk_upload"
 import * as LocalStateStore from "./hooks/local-storage"
 import * as LocalTime from "./hooks/local-time"
 import * as Chat from "./hooks/chat"
@@ -56,13 +56,26 @@ Hooks.ImageForceLoader = ImageForceLoader.hooks
 
 let csrfToken = document.querySelector("meta[name='csrf-token']").getAttribute("content")
 let liveSocket = new LiveSocket("/live", Socket, {
-  params: { _csrf_token: csrfToken, tz_info: Hooks.LocalTime.info() }, hooks: Hooks,
+  params: {
+    _csrf_token: csrfToken,
+    tz_info: Hooks.LocalTime.info(),
+    storage: Hooks.LocalStateStore.fullState({
+      auth_key: "buckitUp-chat-auth-v2",
+      room_count_key: "buckitUp-room-count-v2"
+    })
+  },
+  hooks: Hooks,
   uploaders: Uploaders
 })
 
 const listeners = {
-  "chat:clear-value": (e) => { e.target.value = "" },
-  "chat:focus": (e) => { const el = e.target; setTimeout(() => el.focus(), 100); },
+  "chat:clear-value": (e) => {
+    e.target.value = ""
+  },
+  "chat:focus": (e) => {
+    const el = e.target;
+    setTimeout(() => el.focus(), 100);
+  },
   "chat:toggle": (e) => {
     if (e.detail && e.detail.class) {
       e.target.classList.toggle(e.detail.class)
@@ -83,7 +96,9 @@ const listeners = {
       e.target.style.bottom = '';
     }
 
-    if (relativeElementRect.width < e.target.offsetWidth) { e.target.style.left = 0 }
+    if (relativeElementRect.width < e.target.offsetWidth) {
+      e.target.style.left = 0
+    }
   },
   "chat:select-message": (e) => {
     const messageBlock = e.target;
@@ -93,13 +108,15 @@ const listeners = {
     messageBlockCheckbox.checked = !messageBlockCheckbox.checked;
 
     setTimeout(() => {
-      if (document.querySelector("#chat-messages").classList.contains('selectMode') == false) { return false }
+      if (document.querySelector("#chat-messages").classList.contains('selectMode') == false) {
+        return false
+      }
 
       const allCheckboxes = document.querySelectorAll('.checked')
 
       if (allCheckboxes.length == 0) {
         document.getElementById("chatContent").dispatchEvent(
-          new CustomEvent('chat:toggle-selection-mode', { detail: { chatType: e.detail.chatType } })
+          new CustomEvent('chat:toggle-selection-mode', {detail: {chatType: e.detail.chatType}})
         )
       }
       const deleteButton = document.getElementById("delete-btn");
@@ -137,9 +154,9 @@ const listeners = {
   "phx:chat:scroll": (e) => {
     setTimeout(() => {
       document.querySelector(e.detail.to).scrollIntoView(
-        { behavior: "smooth", block: "center", inline: "nearest" }
+        {behavior: "smooth", block: "center", inline: "nearest"}
       );
-    }, 900) 
+    }, 900)
   },
   "phx:chat:toggle": (e) => {
     if (e.detail && e.detail.class && e.detail.to) {
@@ -153,39 +170,51 @@ const listeners = {
     url = e.detail.url
     url && openUrl(url)
   },
-  "phx:chat:focus": (e) => { const el = document.querySelector(e.detail.to); setTimeout(() => el.focus(), 100); },
-  "phx:chat:change": (e) => { const el = document.querySelector(e.detail.to); el.innerHTML = e.detail.content; },
+  "phx:chat:focus": (e) => {
+    const el = document.querySelector(e.detail.to);
+    setTimeout(() => el.focus(), 100);
+  },
+  "phx:chat:change": (e) => {
+    const el = document.querySelector(e.detail.to);
+    el.innerHTML = e.detail.content;
+  },
   "phx:chat:bulk-change": (e) => {
     const elements = document.querySelectorAll(e.detail.to);
-    elements.forEach((el) => { el.innerHTML = e.detail.content; });
+    elements.forEach((el) => {
+      el.innerHTML = e.detail.content;
+    });
   },
   "phx:scroll-to-bottom": (e) => {
     setTimeout(() => {
       const chatContent = document.querySelector('.a-content-block');
-      chatContent.scrollTo({ top: chatContent.scrollHeight })
+      chatContent.scrollTo({top: chatContent.scrollHeight})
     }, 0)
   },
   "phx:scroll-uploads-to-top": (e) => {
-      const uploader = document.querySelector('.a-uploader');
-      const mobileUploader = document.querySelector('.a-mobile-uploader');
-      uploader.scrollTop = -uploader.scrollHeight;
-      mobileUploader.scrollTop = -mobileUploader.scrollHeight;
+    const uploader = document.querySelector('.a-uploader');
+    const mobileUploader = document.querySelector('.a-mobile-uploader');
+    uploader.scrollTop = -uploader.scrollHeight;
+    mobileUploader.scrollTop = -mobileUploader.scrollHeight;
   },
   "phx:gallery:preload": (e) => {
     const img = new Image();
     img.onload = function () {
       const preloadedList = document.getElementById(e.detail.to);
       preloadedList.appendChild(img);
-      setTimeout(() => { img.remove() }, '30000');
+      setTimeout(() => {
+        img.remove()
+      }, '30000');
     }
     img.classList.add('hidden')
     img.src = e.detail.url;
   },
-  "phx:js-event": (e) => { liveSocket.execJS(document.documentElement, e.detail.data) },
+  "phx:js-event": (e) => {
+    liveSocket.execJS(document.documentElement, e.detail.data)
+  },
   "phx:copy": (e) => {
     navigator.clipboard.writeText(e.target.value)
   },
-  "phx:js-exec": ({ detail }) => {
+  "phx:js-exec": ({detail}) => {
     document.querySelectorAll(detail.to).forEach(el => {
       liveSocket.execJS(el, el.getAttribute(detail.attr))
     })
@@ -197,12 +226,13 @@ for (key in listeners) {
 }
 
 // Show progress bar on live navigation and form submits
-topbar.config({ barColors: { 0: "#29d" }, shadowColor: "rgba(0, 0, 0, .3)" })
+topbar.config({barColors: {0: "#29d"}, shadowColor: "rgba(0, 0, 0, .3)"})
 let topBarScheduled = undefined;
 window.addEventListener("phx:page-loading-start", () => {
   if (!topBarScheduled) {
     topBarScheduled = setTimeout(() => topbar.show(), 120);
-  };
+  }
+  ;
 });
 window.addEventListener("phx:page-loading-stop", () => {
   clearTimeout(topBarScheduled);