+++ lib/chat_web/live/main_live/page/login.ex
@@ -116,6 +116,29 @@ defmodule ChatWeb.MainLive.Page.Login do
   end
 
   def check_stored(socket) do
+    cond do
+      already_loaded_client_storage?(socket) -> socket
+      params = client_storage_in_params(socket) -> emulate_restore_auth(socket, params)
+      true -> request_restore(socket)
+    end
+  end
+
+  defp already_loaded_client_storage?(socket) do
+    match?(%{assigns: %{me: _}}, socket)
+  end
+
+  defp client_storage_in_params(socket) do
+    with %{"storage" => storage} <- Phoenix.LiveView.get_connect_params(socket) do
+      storage
+    end
+  end
+
+  defp emulate_restore_auth(socket, params) do
+    {:noreply, socket} = ChatWeb.MainLive.Index.handle_event("restoreAuth", params, socket)
+    socket |> request_restore()
+  end
+
+  defp request_restore(socket) do
     socket
     |> push_event("restore", %{
       auth_key: @local_store_auth_key,