+++ CHANGELOG.md
@@ -22,6 +22,7 @@ straightforward as possible.
 - Log in by key uploading UI issue [#583]
 - Chats page UI issue [#583]
 - copy invited user invitation with cargo room [#574]
+- relogin from socket state [#538]
 
 ## 2023-12-30_167d8d9___2023-12-30_80fe885